% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sturdy-package.R
\docType{package}
\name{sturdy-package}
\alias{sturdy}
\alias{sturdy-package}
\title{sturdy: Manage A Research Study Database}
\description{
An R package to manage a research study database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mvr-gis.github.io/sturdy/}
}

}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID}) [contributor]
}

}
\keyword{internal}
